# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Commutrum — JSE factor research engine API
servers:
  - url: /api/v1
    description: Versioned API base path
...
paths:
...
  /signals/{ticker}:
    get:
      operationId: getTickerSignals
      tags: [signals]
      summary: Get standardized signals for a ticker
      parameters:
        - $ref: "#/components/parameters/ticker"
        - name: asOf
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Signals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalsResponse"
  /signals/sector-stats:
    get:
      operationId: getSignalSectorStats
      tags: [signals]
      summary: Get cross-sectional signal statistics by sector
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Sector stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectorStatsResponse"

  # ── Module 04 — Orthogonality ────────────────────────────────────────────
...
  /research/{ticker}:
    get:
      operationId: getResearchComposite
      tags: [research]
      summary: Get composite score with factor attribution for a ticker
      parameters:
        - $ref: "#/components/parameters/ticker"
        - name: asOf
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Composite research result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompositeResearchResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /probability/{ticker}:
    get:
      operationId: getTickerProbability
      tags: [probability]
      summary: Get calibrated probability output for a ticker
      parameters:
        - $ref: "#/components/parameters/ticker"
        - name: asOf
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Probability output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProbabilityResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /portfolio/history:
    get:
      operationId: getPortfolioHistory
      tags: [portfolio]
      summary: Get historical portfolio holdings
      parameters:
        - $ref: "#/components/parameters/fromDate"
        - $ref: "#/components/parameters/toDate"
      responses:
        "200":
          description: Portfolio history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioHistoryResponse"
  /portfolio/rebalance:
    post:
      operationId: simulateRebalance
      tags: [portfolio]
      summary: Simulate a portfolio rebalance (does not place real trades)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RebalanceRequest"
      responses:
        "200":
          description: Rebalance simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RebalanceResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ── Module 08 — Execution ────────────────────────────────────────────────
...
  /registry/factors:
    get:
      operationId: getRegistryFactors
      tags: [registry]
      summary: List all factors in the hypothesis registry
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Registry factors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryFactorsResponse"
  /registry/factors/{id}:
    get:
      operationId: getRegistryFactor
      tags: [registry]
      summary: Get full detail for a registry factor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Registry factor detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryFactorDetailResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /registry/audit-log:
    get:
      operationId: getRegistryAuditLog
      tags: [registry]
      summary: Get full audit log
      parameters:
        - name: entityType
          in: query
          required: false
          schema:
            type: string
        - name: entityId
          in: query
          required: false
          schema:
            type: string
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Audit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"
  /registry/factors/{id}/transition:
    post:
      operationId: transitionRegistryFactor
      tags: [registry]
      summary: Transition a factor to a new lifecycle status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifecycleTransitionRequest"
      responses:
        "200":
          description: Transition recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifecycleTransitionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
components:
  parameters:
    ticker:
      name: ticker
      in: path
      required: true
      schema:
        type: string
    fromDate:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date
    toDate:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date
    cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 100
        maximum: 500
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Meta:
      type: object
      properties:
        as_of:
          type: string
          format: date
        generated_at:
          type: string
          format: date-time
      required: [generated_at]
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    SignalEntry:
      type: object
      properties:
        factorId:
          type: string
        rawValue:
          type: number
        sectorRelative:
          type: number
        zScore:
          type: number
        zScoreWinsorized:
          type: number
        sector:
          type: string
        signalDate:
          type: string
          format: date
      required: [factorId, signalDate]
    SignalsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SignalEntry"
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    SectorStatsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              sector:
                type: string
              factorId:
                type: string
              mean:
                type: number
              stdDev:
                type: number
              count:
                type: integer
            required: [sector, factorId]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
...
    CompositeResearchResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            ticker:
              type: string
            scoreDate:
              type: string
              format: date
            compositeScore:
              type: number
            regime:
              type: string
            currencySensitivityFlag:
              type: boolean
            factorAttribution:
              type: object
              additionalProperties:
                type: number
          required: [ticker, scoreDate]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
...
    ProbabilityResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            ticker:
              type: string
            compositeScore:
              type: number
            probabilityPositiveReturn1m:
              type: number
            expectedReturn1m:
              type: number
            confidenceIntervalLow90:
              type: number
            confidenceIntervalHigh90:
              type: number
            historicalWinRate:
              type: number
            calibrationSampleSize:
              type: integer
            riskLevel:
              type: string
          required: [ticker, probabilityPositiveReturn1m]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
...
    Holding:
      type: object
      properties:
        ticker:
          type: string
        weight:
          type: number
        sector:
          type: string
        selectionProbability:
          type: number
        isCash:
          type: boolean
      required: [ticker, weight]
...
    PortfolioHistoryResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              rebalanceDate:
                type: string
                format: date
              holdings:
                type: array
                items:
                  $ref: "#/components/schemas/Holding"
            required: [rebalanceDate, holdings]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    RebalanceRequest:
      type: object
      properties:
        asOf:
          type: string
          format: date
      required: [asOf]
    RebalanceResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            simulationDate:
              type: string
              format: date
            proposedHoldings:
              type: array
              items:
                $ref: "#/components/schemas/Holding"
            estimatedTurnover:
              type: number
            estimatedCost:
              type: number
          required: [simulationDate, proposedHoldings]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
...
    RegistryFactorEntry:
      type: object
      properties:
        factorId:
          type: string
        factorName:
          type: string
        mechanismType:
          type: string
        status:
          type: string
        registeredAt:
          type: string
          format: date-time
      required: [factorId, factorName, status]
    RegistryFactorsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RegistryFactorEntry"
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    RegistryFactorDetailResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            factorId:
              type: string
            factorName:
              type: string
            hypothesis:
              type: string
            mechanismType:
              type: string
            causalityClassification:
              type: string
            dataRequirements:
              type: string
            expectedReturnHorizon:
              type: string
            failureCriteria:
              type: string
            status:
              type: string
            lifecycleHistory:
              type: array
              items:
                type: object
                properties:
                  fromStatus:
                    type: string
                  toStatus:
                    type: string
                  transitionedAt:
                    type: string
                    format: date-time
                  evidence:
                    type: string
                required: [toStatus, transitionedAt]
          required: [factorId, factorName, status]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    AuditLogResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              entityType:
                type: string
              entityId:
                type: string
              action:
                type: string
              performedBy:
                type: string
              performedAt:
                type: string
                format: date-time
            required: [id, entityType, entityId, action, performedAt]
        nextCursor:
          type: string
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]
    LifecycleTransitionRequest:
      type: object
      properties:
        toStatus:
          type: string
          enum: ["R&D", "Validation", "Paper Portfolio", "Production", "Watch", "Retired"]
        evidence:
          type: string
        transitionedBy:
          type: string
        backtestRunId:
          type: string
      required: [toStatus, evidence]
    LifecycleTransitionResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            factorId:
              type: string
            newStatus:
              type: string
          required: [factorId, newStatus]
        meta:
          $ref: "#/components/schemas/Meta"
      required: [data, meta]